[MOB-5009] Extract cached message delete method to utils
