+++ src/inapp/utils.ts
@@ -1,10 +1,10 @@
 import { by } from '@pabra/sortby';
-import { setMany } from 'idb-keyval';
+import { delMany, setMany } from 'idb-keyval';
 import { ANIMATION_DURATION } from 'src/constants';
 import { WebInAppDisplaySettings } from 'src/inapp';
 import { srSpeak } from 'src/utils/srSpeak';
 import { trackInAppDelivery } from '../events';
-import { BrowserStorageEstimate, InAppMessage } from './types';
+import { BrowserStorageEstimate, CachedMessage, InAppMessage } from './types';
 
 interface Breakpoints {
   smMatches: boolean;
@@ -155,6 +155,40 @@ export const determineRemainingStorageQuota = async () => {
   }
   /** do not try to add to cache if we cannot determine storage space */
   return 0;
+};
+
+/**
+ * deletes cached messages not present in latest getMessages fetch
+ * @param cachedMessages
+ * @param fetchedMessages
+ */
+export const deleteMessagesFromCache = async (
+  cachedMessages: CachedMessage[],
+  fetchedMessages: Partial<InAppMessage>[]
+) => {
+  const cachedMessagesToDelete = cachedMessages.reduce(
+    (deleteQueue: string[], [cachedMessageId]) => {
+      const isCachedMessageInFetch = fetchedMessages.reduce(
+        (isFound, { messageId }) => {
+          if (messageId === cachedMessageId) isFound = true;
+          return isFound;
+        },
+        false
+      );
+
+      if (!isCachedMessageInFetch) deleteQueue.push(cachedMessageId);
+      return deleteQueue;
+    },
+    []
+  );
+  try {
+    await delMany(cachedMessagesToDelete);
+  } catch (err: any) {
+    console.warn(
+      'Error deleting messages from the browser cache',
+      err?.response?.data?.clientErrors ?? err
+    );
+  }
 };
 
 /**
